feat: Add functionality to add and display books in the library
The code changes include adding functions to add books to the library and display all the books in the library. This will allow users to easily add new books and view the existing books in the library.

Note: The recent user and repository commits do not provide any additional relevant information for the commit message.